OAuth: Move the oauth reducer to the UI state tree
