+++ client/blocks/login/index.jsx
@@ -21,7 +21,7 @@ import {
 	getSocialAccountIsLinking,
 	getSocialAccountLinkService,
 } from 'state/login/selectors';
-import { getOAuth2ClientData } from 'state/login/oauth2/selectors';
+import { getOAuth2ClientData } from 'state/ui/oauth2-clients/selectors';
 import { recordTracksEvent } from 'state/analytics/actions';
 import VerificationCodeForm from './two-factor-authentication/verification-code-form';
 import WaitingTwoFactorNotificationApproval from './two-factor-authentication/waiting-notification-approval';
